+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [ ] refuse from stateful `Imohash` class in favor of two functions with arguments instead
 - [ ] add support for `typing.IO` objects instead of too narrow file path
 - [ ] move multithreading logic from Python to Rust
+- [x] cover with unit-tests
 
 ## [0.1.0] — 2025-04-16